Done with the docker and othe stuff
