+++ application/.docker.env
@@ -1,5 +1,5 @@
 # Production information.
-MYSQL_HOST=
+MYSQL_HOST=application_db
 MYSQL_DATABASE=application
 MYSQL_USER=root
 MYSQL_PASSWORD=root
