+++ docker-compose.override.yml
@@ -8,8 +8,13 @@ services:
   application:
     volumes:
       - ./application:/home/app
-      - /home/app/node_modules
-    command: ["npm", "run", "dev"]
+    ports:
+      - '3000:3000'
+
+  application_db:
+    restart: always
+    ports:
+      - '3306:3306'
 
   gateway:
     volumes:
